extract dns resolver code to /discovery package
