+++ pkg/adscert/discovery/resolver_api.go
@@ -1,0 +1,7 @@
+package discovery
+
+import "context"
+
+type DNSResolver interface {
+	LookupTXT(ctx context.Context, name string) ([]string, error)
+}
